fedora-langpacks: Skip self tests under non-fedora distros
The fedora-langpacks plugin disables itself if not on fedora. Skip the
self tests in that situation as well so they don't fail. The reason this
hasn't caused CI failures is that we have "--no-suite fedora-langpacks"
in .gitlab-ci.yml so that it doesn't run under Debian. But it still
seems worth skipping as people may run the tests manually, and we need
to skip it in the downstream Endless CI.