gs-plugin-loader: Skip saving install queue during unit tests
Currently some unit tests fail with a message like:
Bail out! Gs-FATAL-WARNING: failed to create an app for */*/*/*/chiron.desktop/*

This is because the tests are being run in parallel, and in
gs_plugin_loader_setup() we load the install queue from the disk, but
the queue is also being written to by other tests so it may have apps in
it which can't be understood by the plugin being tested (chiron.desktop
is only used by the dummy plugin).

So avoid saving the install queue to disk during the unit tests.